Update file NIO support:
 - Create Pull for FileHandle from either paths or channels
 - Move Pull constructors to fs2.io.pulls
 - Move FileHandle constructors to fs2.io.FileHandle companion
 - Make readAllFromFileHandle/writeAllToFileHandle visible
